feat: rearranged folder structure for diffsync
Rearranged the folder structure so that the differential synchronization
algorithm is isolated. This was done since new synchronization
algorithms will be implemented in the near future. On top of that,
the main classes Client and Server were renamed to OTClient and
OTServer, respectively to namespace them from the client and server
classes for other algorithms.

BREAKING CHANGE: Client and Server renamed, folder structure changed